# Backend_Jobaile/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Jobaile API",
        "description": "Dokumentasi API untuk aplikasi Jobaile (versi pertama)",
        "contact": {
            "name": "Jobaile Dev Team",
            "email": "support@jobaile.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Server utama"
        }
    ],
    "paths": {
...
        "/api/payments/qr/{contractId}": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Generate QR Payment untuk kontrak",
                "description": "`* @OA\\Post(\n    path=\"/api/payments/qr/{contractId}\",\n    summary=\"Generate QR Payment untuk kontrak\",\n    tags={\"Payment\"},\n    security={{\"sanctum\":{}}},",
                "operationId": "0990b4e8366d461779f38b91333a5947",
                "parameters": [
                    {
                        "name": "contractId",
                        "in": "path",
                        "description": "ID dari kontrak",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "QR code berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "QR code berhasil dibuat."
                                        },
                                        "data": {
                                            "properties": {
                                                "qr_code_url": {
                                                    "type": "string",
                                                    "example": "https://dummy.dana.qr/scan/JOB-uuid123"
                                                },
                                                "payment_id": {
                                                    "type": "integer",
                                                    "example": 101
                                                },
                                                "merchant_trans_id": {
                                                    "type": "string",
                                                    "example": "JOB-uuid123"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "waiting_for_payment"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User belum terautentikasi"
                    },
                    "403": {
                        "description": "Email belum diverifikasi atau bukan recruiter"
                    },
                    "404": {
                        "description": "Kontrak tidak ditemukan"
                    },
                    "400": {
                        "description": "Kontrak sudah dibayar"
                    },
                    "500": {
                        "description": "Kesalahan server"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/payments/callback": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Callback DANA untuk update status pembayaran",
                "operationId": "2e7a00bb7d2efc05e03867b2cc5b135e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "merchantTransId",
                                    "status"
                                ],
                                "properties": {
                                    "merchantTransId": {
                                        "type": "string",
                                        "example": "JOB-uuid123"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "success",
                                            "failed",
                                            "pending"
                                        ],
                                        "example": "success"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status pembayaran diperbarui"
                    },
                    "404": {
                        "description": "Pembayaran tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
...
    },
...
}